Hungarian Algorithm
Hungarian Algorithm: This project is written in C++. The other name of the project is "Matching/Assignment Problem". Since the algorithm is well designed, it works effectively and smoothly not only in small sized (3x3, 4x4, etc.) matrices, but also in large sized matrices. The program works on a console basis. Information about the application, the working logic of the system and more are included in the report document.